pyproject.toml

[tool.poetry]
name = "spalah"
version = "0.5.0"
description = "Spalah is a set of PySpark dataframe helpers"
homepage = "https://github.com/avolok/spalah"
authors = ["Alex Volok <alexandr.volok@gmail.com>"]
license = "MIT license"
keywords = ["spalah"]
classifiers = ["Intended Audience :: Developers", "License :: OSI Approved :: MIT License", "Natural Language :: English", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.7", "Programming Language :: Python :: 3.8", "Programming Language :: Python :: 3.9", "Programming Language :: Python :: 3.10"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0.0"
black = "==22.3.0"
bump2version = "==1.0.1"
coverage = "==6.4.1"
delta-spark = "==2.1.0"
flake8 = "==6.0.0"
packaging = "*"
pip = "*"
pre-commit = "==2.19.0"
pyspark = "==3.3.0"
pytest = "==7.1.2"
pytest-cov = "==3.0.0"
ruff = "*"
tox = "==3.25.0"
twine = "==4.0.1"
watchdog = "==2.1.9"
wheel = "*"


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.1.13"
mkdocstrings = {extras = ["python"], version = "^0.20.0"}
mkdocs = ">=1.4.3"
termynal = ">=0.2.1"

[tool.semantic_release]
version_variable = [    
    "pyproject.toml:version"
]
build_command = "pip install poetry && poetry build"
branch = "main"

[tool.ruff]
line-length = 100
fix = true